Moved amcrest.py to __init__.py to easy the imports
